chore(suite): moved * simplified useRefetchFees into wallet-core
