+++ suite-common/wallet-core/package.json
@@ -38,6 +38,7 @@
         "date-fns": "^4.1.0",
         "react": "18.2.0",
         "react-hook-form": "^7.59.0",
+        "react-redux": "^9.2.0",
         "web3-utils": "^4.3.1"
     }
 }
